+++ tests/index.js
@@ -146,7 +146,7 @@ async function go(showData) {
   // logger() begin
 
   // eslint-disable-next-line no-self-compare
-  if (loggers.logger('dog') !== loggers.logger('dog')) throw new Error();
+  if (loggers.logger('dog') !== loggers.logger('dog')) throw new Error(); // test caching
 
   // 'dog' doesn't log at silly
   {
@@ -158,6 +158,7 @@ async function go(showData) {
   // Check disabling the warn tag
   if (loggers.isLevelEnabled({ tags: { silly: true, warn: false }, category: 'dog' })) throw new Error();
 
+  // Cache 'dog' child logger
   // 'dog' will log at silly because warn is added
   loggers.logger('dog', loggers.child(['warn', 'goofy'], { dog: 'woof', a: {b: {d: 5}} }));