+++ changelog/CHANGELOG.md
@@ -2,13 +2,17 @@
 
 ## 8.0.0 2022-06-20
 
-Breaking changes
+### Breaking changes
 
 - Add 'context'
   - All forms of log functions accept a context
   - Child loggers use context instead of data
 - 'data' event renamed to 'redact'
 - data() method removed from Loggers and Logger classes
+
+### Enhancements
+
+Context, message, and data objects are sent to 'redact' event separately
 
 ## 7.0.1 2022-05-10